+++ contract/contracts/distributors/WalletNFT.sol
@@ -48,6 +48,6 @@ contract WalletNFTDistributor is NFTDistributorInterfaceV1, MerkleTreeManager {
     ) virtual external {
         require(proof(treeId, index, account, 1, merkleProof));
 
-        _mint(account, uint256(treeId), 1, "");
+        _mint(account, treeId, 1, "");
     }
 }
